+++ src/org/magnos/impulse/CollisionPolygonPolygon.java
@@ -21,31 +21,27 @@
 
 package org.magnos.impulse;
 
-public class CollisionPolygonPolygon implements CollisionCallback
-{
+public class CollisionPolygonPolygon implements CollisionCallback {
 
 	public static final CollisionPolygonPolygon instance = new CollisionPolygonPolygon();
 
 	@Override
-	public void handleCollision( Manifold m, Body a, Body b )
-	{
-		Polygon A = (Polygon)a.shape;
-		Polygon B = (Polygon)b.shape;
+	public void handleCollision(Manifold m, Body a, Body b) {
+		Polygon A = (Polygon) a.shape;
+		Polygon B = (Polygon) b.shape;
 		m.contactCount = 0;
 
 		// Check for a separating axis with A's face planes
 		int[] faceA = { 0 };
-		float penetrationA = findAxisLeastPenetration( faceA, A, B );
-		if (penetrationA >= 0.0f)
-		{
+		double penetrationA = findAxisLeastPenetration(faceA, A, B);
+		if (penetrationA >= 0.0f) {
 			return;
 		}
 
 		// Check for a separating axis with B's face planes
 		int[] faceB = { 0 };
-		float penetrationB = findAxisLeastPenetration( faceB, B, A );
-		if (penetrationB >= 0.0f)
-		{
+		double penetrationB = findAxisLeastPenetration(faceB, B, A);
+		if (penetrationB >= 0.0f) {
 			return;
 		}
 
@@ -56,15 +52,12 @@ public class CollisionPolygonPolygon implements CollisionCallback
 		Polygon IncPoly; // Incident
 
 		// Determine which shape contains reference face
-		if (ImpulseMath.gt( penetrationA, penetrationB ))
-		{
+		if (ImpulseMath.gt(penetrationA, penetrationB)) {
 			RefPoly = A;
 			IncPoly = B;
 			referenceIndex = faceA[0];
 			flip = false;
-		}
-		else
-		{
+		} else {
 			RefPoly = B;
 			IncPoly = A;
 			referenceIndex = faceB[0];
@@ -72,9 +65,9 @@ public class CollisionPolygonPolygon implements CollisionCallback
 		}
 
 		// World space incident face
-		Vec2[] incidentFace = Vec2.arrayOf( 2 );
+		Vec2[] incidentFace = Vec2.arrayOf(2);
 
-		findIncidentFace( incidentFace, RefPoly, IncPoly, referenceIndex );
+		findIncidentFace(incidentFace, RefPoly, IncPoly, referenceIndex);
 
 		// y
 		// ^ .n ^
@@ -97,69 +90,62 @@ public class CollisionPolygonPolygon implements CollisionCallback
 		// Transform vertices to world space
 		// v1 = RefPoly->u * v1 + RefPoly->body->position;
 		// v2 = RefPoly->u * v2 + RefPoly->body->position;
-		v1 = RefPoly.u.mul( v1 ).addi( RefPoly.body.position );
-		v2 = RefPoly.u.mul( v2 ).addi( RefPoly.body.position );
+		v1 = RefPoly.u.mul(v1).addi(RefPoly.body.position);
+		v2 = RefPoly.u.mul(v2).addi(RefPoly.body.position);
 
 		// Calculate reference face side normal in world space
 		// Vec2 sidePlaneNormal = (v2 - v1);
 		// sidePlaneNormal.Normalize( );
-		Vec2 sidePlaneNormal = v2.sub( v1 );
+		Vec2 sidePlaneNormal = v2.sub(v1);
 		sidePlaneNormal.normalize();
 
 		// Orthogonalize
 		// Vec2 refFaceNormal( sidePlaneNormal.y, -sidePlaneNormal.x );
-		Vec2 refFaceNormal = new Vec2( sidePlaneNormal.y, -sidePlaneNormal.x );
+		Vec2 refFaceNormal = new Vec2(sidePlaneNormal.y, -sidePlaneNormal.x);
 
 		// ax + by = c
 		// c is distance from origin
 		// real refC = Dot( refFaceNormal, v1 );
 		// real negSide = -Dot( sidePlaneNormal, v1 );
 		// real posSide = Dot( sidePlaneNormal, v2 );
-		float refC = Vec2.dot( refFaceNormal, v1 );
-		float negSide = -Vec2.dot( sidePlaneNormal, v1 );
-		float posSide = Vec2.dot( sidePlaneNormal, v2 );
+		double refC = Vec2.dot(refFaceNormal, v1);
+		double negSide = -Vec2.dot(sidePlaneNormal, v1);
+		double posSide = Vec2.dot(sidePlaneNormal, v2);
 
 		// Clip incident face to reference face side planes
 		// if(Clip( -sidePlaneNormal, negSide, incidentFace ) < 2)
-		if (clip( sidePlaneNormal.neg(), negSide, incidentFace ) < 2)
-		{
+		if (clip(sidePlaneNormal.neg(), negSide, incidentFace) < 2) {
 			return; // Due to floating point error, possible to not have required
-						// points
+					// points
 		}
 
 		// if(Clip( sidePlaneNormal, posSide, incidentFace ) < 2)
-		if (clip( sidePlaneNormal, posSide, incidentFace ) < 2)
-		{
+		if (clip(sidePlaneNormal, posSide, incidentFace) < 2) {
 			return; // Due to floating point error, possible to not have required
-						// points
+					// points
 		}
 
 		// Flip
-		m.normal.set( refFaceNormal );
-		if (flip)
-		{
+		m.normal.set(refFaceNormal);
+		if (flip) {
 			m.normal.negi();
 		}
 
 		// Keep points behind reference face
 		int cp = 0; // clipped points behind reference face
-		float separation = Vec2.dot( refFaceNormal, incidentFace[0] ) - refC;
-		if (separation <= 0.0f)
-		{
-			m.contacts[cp].set( incidentFace[0] );
+		double separation = Vec2.dot(refFaceNormal, incidentFace[0]) - refC;
+		if (separation <= 0.0f) {
+			m.contacts[cp].set(incidentFace[0]);
 			m.penetration = -separation;
 			++cp;
-		}
-		else
-		{
+		} else {
 			m.penetration = 0;
 		}
 
-		separation = Vec2.dot( refFaceNormal, incidentFace[1] ) - refC;
+		separation = Vec2.dot(refFaceNormal, incidentFace[1]) - refC;
 
-		if (separation <= 0.0f)
-		{
-			m.contacts[cp].set( incidentFace[1] );
+		if (separation <= 0.0f) {
+			m.contacts[cp].set(incidentFace[1]);
 
 			m.penetration += -separation;
 			++cp;
@@ -171,27 +157,25 @@ public class CollisionPolygonPolygon implements CollisionCallback
 		m.contactCount = cp;
 	}
 
-	public float findAxisLeastPenetration( int[] faceIndex, Polygon A, Polygon B )
-	{
-		float bestDistance = -Float.MAX_VALUE;
+	public double findAxisLeastPenetration(int[] faceIndex, Polygon A, Polygon B) {
+		double bestDistance = -Double.MAX_VALUE;
 		int bestIndex = 0;
 
-		for (int i = 0; i < A.vertexCount; ++i)
-		{
+		for (int i = 0; i < A.vertexCount; ++i) {
 			// Retrieve a face normal from A
 			// Vec2 n = A->m_normals[i];
 			// Vec2 nw = A->u * n;
-			Vec2 nw = A.u.mul( A.normals[i] );
+			Vec2 nw = A.u.mul(A.normals[i]);
 
 			// Transform face normal into B's model space
 			// Mat2 buT = B->u.Transpose( );
 			// n = buT * nw;
 			Mat2 buT = B.u.transpose();
-			Vec2 n = buT.mul( nw );
+			Vec2 n = buT.mul(nw);
 
 			// Retrieve support point from B along -n
 			// Vec2 s = B->GetSupport( -n );
-			Vec2 s = B.getSupport( n.neg() );
+			Vec2 s = B.getSupport(n.neg());
 
 			// Retrieve vertex on face from A, transform into
 			// B's model space
@@ -199,15 +183,14 @@ public class CollisionPolygonPolygon implements CollisionCallback
 			// v = A->u * v + A->body->position;
 			// v -= B->body->position;
 			// v = buT * v;
-			Vec2 v = buT.muli( A.u.mul( A.vertices[i] ).addi( A.body.position ).subi( B.body.position ) );
+			Vec2 v = buT.muli(A.u.mul(A.vertices[i]).addi(A.body.position).subi(B.body.position));
 
 			// Compute penetration distance (in B's model space)
 			// real d = Dot( n, s - v );
-			float d = Vec2.dot( n, s.sub( v ) );
+			double d = Vec2.dot(n, s.sub(v));
 
 			// Store greatest distance
-			if (d > bestDistance)
-			{
+			if (d > bestDistance) {
 				bestDistance = d;
 				bestIndex = i;
 			}
@@ -217,30 +200,27 @@ public class CollisionPolygonPolygon implements CollisionCallback
 		return bestDistance;
 	}
 
-	public void findIncidentFace( Vec2[] v, Polygon RefPoly, Polygon IncPoly, int referenceIndex )
-	{
+	public void findIncidentFace(Vec2[] v, Polygon RefPoly, Polygon IncPoly, int referenceIndex) {
 		Vec2 referenceNormal = RefPoly.normals[referenceIndex];
 
 		// Calculate normal in incident's frame of reference
 		// referenceNormal = RefPoly->u * referenceNormal; // To world space
 		// referenceNormal = IncPoly->u.Transpose( ) * referenceNormal; // To
 		// incident's model space
-		referenceNormal = RefPoly.u.mul( referenceNormal ); // To world space
-		referenceNormal = IncPoly.u.transpose().mul( referenceNormal ); // To
-																								// incident's
-																								// model
-																								// space
+		referenceNormal = RefPoly.u.mul(referenceNormal); // To world space
+		referenceNormal = IncPoly.u.transpose().mul(referenceNormal); // To
+																		// incident's
+																		// model
+																		// space
 
 		// Find most anti-normal face on incident polygon
 		int incidentFace = 0;
-		float minDot = Float.MAX_VALUE;
-		for (int i = 0; i < IncPoly.vertexCount; ++i)
-		{
+		double minDot = Double.MAX_VALUE;
+		for (int i = 0; i < IncPoly.vertexCount; ++i) {
 			// real dot = Dot( referenceNormal, IncPoly->m_normals[i] );
-			float dot = Vec2.dot( referenceNormal, IncPoly.normals[i] );
+			double dot = Vec2.dot(referenceNormal, IncPoly.normals[i]);
 
-			if (dot < minDot)
-			{
+			if (dot < minDot) {
 				minDot = dot;
 				incidentFace = i;
 			}
@@ -254,31 +234,29 @@ public class CollisionPolygonPolygon implements CollisionCallback
 		// v[1] = IncPoly->u * IncPoly->m_vertices[incidentFace] +
 		// IncPoly->body->position;
 
-		v[0] = IncPoly.u.mul( IncPoly.vertices[incidentFace] ).addi( IncPoly.body.position );
-		incidentFace = incidentFace + 1 >= (int)IncPoly.vertexCount ? 0 : incidentFace + 1;
-		v[1] = IncPoly.u.mul( IncPoly.vertices[incidentFace] ).addi( IncPoly.body.position );
+		v[0] = IncPoly.u.mul(IncPoly.vertices[incidentFace]).addi(IncPoly.body.position);
+		incidentFace = incidentFace + 1 >= IncPoly.vertexCount ? 0 : incidentFace + 1;
+		v[1] = IncPoly.u.mul(IncPoly.vertices[incidentFace]).addi(IncPoly.body.position);
 	}
 
-	public int clip( Vec2 n, float c, Vec2[] face )
-	{
+	public int clip(Vec2 n, double c, Vec2[] face) {
 		int sp = 0;
-		Vec2[] out = {
-			new Vec2( face[0] ),
-			new Vec2( face[1] )
-		};
+		Vec2[] out = { new Vec2(face[0]), new Vec2(face[1]) };
 
 		// Retrieve distances from each endpoint to the line
 		// d = ax + by - c
 		// real d1 = Dot( n, face[0] ) - c;
 		// real d2 = Dot( n, face[1] ) - c;
-		float d1 = Vec2.dot( n, face[0] ) - c;
-		float d2 = Vec2.dot( n, face[1] ) - c;
+		double d1 = Vec2.dot(n, face[0]) - c;
+		double d2 = Vec2.dot(n, face[1]) - c;
 
 		// If negative (behind plane) clip
 		// if(d1 <= 0.0f) out[sp++] = face[0];
 		// if(d2 <= 0.0f) out[sp++] = face[1];
-		if (d1 <= 0.0f) out[sp++].set( face[0] );
-		if (d2 <= 0.0f) out[sp++].set( face[1] );
+		if (d1 <= 0.0f)
+			out[sp++].set(face[0]);
+		if (d2 <= 0.0f)
+			out[sp++].set(face[1]);
 
 		// If the points are on different sides of the plane
 		if (d1 * d2 < 0.0f) // less than to ignore -0.0f
@@ -288,9 +266,9 @@ public class CollisionPolygonPolygon implements CollisionCallback
 			// out[sp] = face[0] + alpha * (face[1] - face[0]);
 			// ++sp;
 
-			float alpha = d1 / (d1 - d2);
+			double alpha = d1 / (d1 - d2);
 
-			out[sp++].set( face[1] ).subi( face[0] ).muli( alpha ).addi( face[0] );
+			out[sp++].set(face[1]).subi(face[0]).muli(alpha).addi(face[0]);
 		}
 
 		// Assign our new converted values
